NPCTool_1_1_02_BetaPack zipped folder
The new tool comes in a zipped set of folders that also include the manifest.json files to allow simply placing into the appropriate development folders for minecraft bedrock or minecraft education.
The main excel workbook tool is now just an .xlsm file instead of a template file as that was causing issues with getting the save copy macros to work properly in the correct folder.
There is an instruction sheet in the workbook.  Copy and rename the NPCToolTest_blank sheets so you can fill in with your own material.  Make sure your copy of the workbook file is saved to the folder where you want to find the results.  Then hit the Fill formula and Make Files button on the Top Left of the Sheet you are ready to create the dialogue and language files for.
This is definitely still a Beta so I'm sure there are upgrades, issues and bugs I should fix.  Please let me know your findings so I can make adjustments to this.